refactor: split func
