+++ packages/scaffold/src/core/builder.ts
@@ -14,6 +14,7 @@ import { glob } from "tinyglobby";
 import { generateHash } from "../utils/crypto.js";
 import { dateFormat, replaceInFile, toArray } from "../utils/string.js";
 import { Base } from "./base.js";
+import { parsePrefs, renderDts } from "./builders/prefs.js";
 
 export default class Build extends Base {
   private buildTime: string;
@@ -248,62 +249,14 @@ export default class Build extends Base {
       return;
 
     const prefsContent = await readFile(prefsFilePath, "utf-8");
-    // eslint-disable-next-line regexp/no-super-linear-backtracking
-    const prefPattern = /^pref\s*\(\s*["']([^"']+)["']\s*,\s*(.+)\s*\)\s*;$/gm;
-    const prefsMap: Map<string, string | boolean | number > = new Map();
-
-    const matches = prefsContent.matchAll(prefPattern);
-    for (const match of matches) {
-      const key = match[1].trim();
-      const value = match[2].trim();
-
-      let parsedValue: string | boolean | number;
-      if (value === "true" || value === "false") {
-        parsedValue = value === "true";
-      }
-      else if (/^\d+$/.test(value)) {
-        parsedValue = Number(value);
-      }
-      else {
-        parsedValue = value.replace(/['"]/g, "");
-      }
-
-      if (prefsMap.has(key))
-        this.logger.warn(`Duplicate preference keys found: ${key}`);
-      else
-        prefsMap.set(key, parsedValue);
-    };
+    const prefsMap = parsePrefs(prefsContent);
 
     if (this.ctx.build.prefs.prefixPrefKeys) {
       //
     }
 
     if (this.ctx.build.prefs.dts) {
-      const dtsContent = `
-/* prettier-ignore */
-/* eslint-disable */
-// @ts-nocheck
-// Generated by zotero-plugin-scaffold
-
-type _PluginPrefsMap = {
-  ${Array.from(prefsMap)
-    .map(pref => `"${pref[0].replace(this.ctx.build.prefs.prefix, "")}": ${typeof pref[1]};`)
-    .join("\n  ")}
-};
-
-type PluginPrefKey<K extends keyof _PluginPrefsMap> = \`${this.ctx.build.prefs.prefix}\${K}\`;
-
-type PluginPrefsMap = {
-  [K in keyof _PluginPrefsMap as PluginPrefKey<K>]: _PluginPrefsMap[K]
-};
-
-declare namespace _ZoteroTypes {
-  interface Prefs {
-    get: <K extends keyof PluginPrefsMap>(key: K, global?: boolean) => PluginPrefsMap[K];
-    set: <K extends keyof PluginPrefsMap>(key: K, value: PluginPrefsMap[K], global?: boolean) => any;
-  }
-}
-`;
+      const dtsContent = renderDts(prefsMap, this.ctx.build.prefs.prefix);
 
       let dtsFilePath = `typings/prefs.d.ts`;
       if (typeof this.ctx.build.prefs.dts === "string")
